Add Synchronizer class for basic sync functions
For our first pass at this, we’re just throwing every record on 
Salesforce into the DB blindly. This is obviously not a tenable long-
term solution, so the next step is to prevent the creation of new
records when an existing record is already associated with the record in
Salesforce.