Normalize `Mapping#attributes` hashes
It’s a lot easier to deal with attribute hashes if they’re passed around
in a consistent format. This commit ensures that `#attributes` dumps
a standard Hash which `#convert` can then massage into any form.

This is a more flexible approach that would allow for creating/copying 
records of the same type (though that sort of sync flexibility is not a
goal in the short term).