Add ActiveRecord model integration point
In order to add Restforce::DB’s functionality to a suite of ActiveRecord
models trivially, we need to establish a trivial entry point into the 
system. In our case, it makes sense to register new RecordTypes through
a module that developers can freely attach to their models.

Now, including Restforce::DB::Module on a class provides a lightweight
DSL for mapping the class to a record type in Salesforce.