Clean up logging output
The current logging format is unnecessarily verbose — for instance, the
default Logger output already includes a timestamp, meaning that our 
logs end up with two timestamps.

This commit moves to a much simpler log style, including just the
log level, a single timestamp, and the logged text, for less noise.