Add synchronized_at timestamp to track sync times
To prepare for future performance and query optimizations, we need some
sort of mechanism to track the times of the most recent synchronization
between a database record and its corresponding Salesforce record.

The approach taken in this commit is to update a timestamp column which
is expected to exist for all models implementing Restforce::DB’s
functionality. This will allow us to short-circuit a lot of processing
based on a comparison of the sync and update timestamps for our records.