Track synchronization runtimes for run efficiency
It’s usually highly undesirable to run the synchronization loop for all
data since the beginning of time. This commit adds a `Tracker` utility 
class which populates a global `last_run` attribute from a supplied
tracking file, and records the timestamps of successful runs into the 
tracking file.

This allows subsequently-created Workers to start from a time of known-
synchronized data, and can be used to configure an initial starting 
point for integration testing purposes.