Consolidate RecordType into Mapping
To expand on our synchronization functionality in the future, there does
not seem to be a whole lot of use for a class that just manages the
manipulation of attribute hashes. In the interest of keeping the class
hierarchy fairly flat, the functionality previously encompassed by
RecordType has been rolled in to the Mapping class.

This is the first step toward supporting the automatic creation of 
associated records, given an entry point into the application.