Prevent record creation for non-root mappings
It may be desirable to exclude specific records from being added to the 
database directly; e.g., the Salesforce model is multipurpose, and we 
only want to synchronize those records associated with a different 
record type that we _do_ care about.

To that end, this commit supports a `:root` option for Mapping objects,
which defines whether or not we want to sync all records without running
through any other filter criteria. This will allow us to introduce and
handle record associations in a future commit.