Ensure that worker configuration loads properly
In eager-loaded applications, there’s a possibility of Restforce::DB’s
configuration being loaded prematurely, which results in a lot of 
useless errors when the daemon runs through its polling loop.

To hopefully prevent these issues, we’re resetting the configuration
just before we parse the config file in the worker.