Support arbitrary lookup constraints for records
There are many cases where you might want to restrict the set of 
returned records from Salesforce to adhere to some sort of business
logic. This commit adds a `conditions` Array to `Mapping`, which defines
a set of lookup conditions that are applied to all Salesforce lookups
for the registered mapping.