Delay Synchronizer queries to account for drift
If Salesforce's server time happens to be ahead of our local server 
time, there's no real issue; we'll eventually ingest changed records 
even if we get to them a few seconds late. If, however, Salesforce's 
server is behind ours by even a few milliseconds, we risk missing 
records which were updated within the offset window.

To combat this, we introduce a slight delay into our queries, looking
for updates some number of seconds/milliseconds in the past, to give
Salesforce a chance to “catch up”.