Add support for explicit Salesforce ID columns
As a step toward supporting multiple mappings on a single model, we need
to support the storage of multiple Salesforce IDs in a single database
table. The approach used here is to lean on an explicit column when one
is present, and falling back to the generic `salesforce_id` column when
one is not.