Support multiple mappings for the same model
On occasion, we may need to support one-to-several relationships in our
mappings, for database records composed of information across multiple
Salesforce object types. In these cases, we need to store more than one
mapping for the database record, and ensure that associations are 
populated with data from all of their mappings.

This commit modifies the current “root” definition; rather than forcing
models to opt in to top-level synchronization, we will instead allow
models to opt _out_ of it by passing a `through` option to `sync_with`.
When this option is present, we know that the mappings should be 
initiated from the creation callbacks of another mapping.