Write first pass at an “Accumulator” object
To keep our synchronization runs internally consistent, we need to
aggregate changes over the course of a run, then apply our changes to
all objects once we’ve determined which changes are most recent. In 
order to pull this off, we need a way of aggregating changes by their
modification timestamps.