Introduce Collector object for field accumulation
To work in tandem with our new Accumulator class, we’re introducing a 
Collector object, which will take over a role currently filled by the
Synchronizer class: iterating through recently-changed records to 
determine which changes to persist to their mapped records.

It builds a generic Hash by default, but a custom accumulator can be 
injected as an option (Accumulator’s API is Hash-like for this very 
purpose).