Refactor timing concerns into a Runner object
As we begin to break apart the Synchronizer class into smaller pieces,
we have an increased need to abstract out our “recent record” lookup
functionality. To that end, we’ve extracted the functionality into
Restforce::DB::Runner, which exposes an API to manage the timestamps 
used for a set of lookup queries.