Factor object creation pattern into Initializer
As the next step of ultimately replacing the existing Synchronizer
object, we’re extracting out the object creation step into its own 
object. Restforce::DB::Initializer is responsible for ensuring that each
record in Salesforce or the database ends up with a matching record in
the other system, prior to synchronizing updated data.