Prevent early `Runner` instantiation
When a tracker is supplied to our worker object, we need to ensure that
its `last_run` configuration is respected by the `Runner` we use for our
workflows. To make this happen reliably, we need to lean on some lazy
instantiation.

Prior to this fix, we were seeing unscoped queries on the first pass of
our daemon, despite the tracker loading in a recent timestamp.