Refactor initialization using Strategy pattern
In order to support has-many relationships (where the lookup ID is on 
the associated record and not the base record), we need to allow for 
more robust decision-making around when objects should be added to each
system. The first step to this goal is to refactor the current logic in
Initializer to delegate the decisions to drop-in strategy objects.

We’ll add new strategies as we go.