Define a custom DSL for configuring model Mappings
To make our mapping/synchronization setup more intuitive in the context
of a project using ActiveRecord, we’ve defined our own minimal DSL for
creating and configuring a record type Mapping.

Currently, this DSL is a pure replacement for the previously-handled
`sync_with` options, but it gives us the freedom to build out new 
association types, and should result in less API churn over time.