Recursively build nested associations
Currently, we only build a single layer of associated records. In order
to support more complicated relationships between records, we need to
allow for the creation of associated records at arbitrary depths.

The Association subclasses already give us most of the functionality we
need, but we need to recursively iterate over the associations for 
nested mappings while building out each individual association.