Prevent the construction of invalid associations
When an associated record is filtered out (either because it simply
doesn’t exist, or, more likely, because it doesn’t meet the criteria
imposed by the `where` clauses for a specific mapping), we should avoid
attempting to construct the association.

In these cases, it’s up to the developer building out the integration to
ensure that the record can be later synchronized — for example, by using
an `associated` strategy for the mapping which will pick the record up
on later modification.