Clean up change accumulation process
Currently, if a Salesforce object type is shared across multiple 
ActiveRecord models, the only changeset which is respected is the one
corresponding to the most recently-registered mapping. This means that
several other mappings may end up with empty diffs, despite relevant 
changes to their data actually having been made in Salesforce.

The solution is to build up a master changeset across all mapping 
entries for the same timestamp.