Prevent the recreation of any pre-existing records
Our current approach to the construction of associated records uses a
flawed approach to looking up existing records through a lookup Hash.
Because the association will _not yet_ be associated to the record under
construction, using the scope from the association will _never_ identify
the appropriate record, even if one with matching Salesforce IDs exists.

The solution is to query directly through the ActiveRecord class, and
assign the record after a match has been found. We can also take this
opportunity to short-circuit the building of additional associated
records, as we can assume any previously-synchronized record is already
being updated appropriately.