Add rake task to seed entire record categories
Once Restforce::DB is up and running with a given mapping, it will stay
up to date over the course of each synchronization. However, when 
setting up an initial mapping, it’s often desirable to synchronize all
existing information.

Currently, this requires manually updating the individual Salesforce
records to force a sync. To obviate the need for such a tedious manual 
process, we’re adding a “seed” Rake task to perform this initial record
population.