Allow access to “unscoped” Salesforce queries
In order to detect records which should be dropped from the database, we
need to be able to compare the list of all modified records for a time
span with all _valid_ records for that same time span. The missing 
ingredient at the moment is the ability to grab the entire collection,
given that `#each` automatically applies the Mapping conditions.