Add AssociationCache to track records being built
In order to handle circular associations (which would likely result in
an infinite association-building loop), we need to be able to detect
when we’re trying to construct a record we’ve already constructed. To
that end, we’re introducing an AssociationCache object, which is
responsible for fetching pre-existing association records.