Pass AssociationCache to nested association builds
In order to ensure that association records aren’t reconstructed for any
specific record, we need to maintain a single AssociationCache through
each of our calls to `#build` for nested associations.

The approach taken here relies on storing the cache in an instance 
variable, which is likely thread-unsafe, but the worker daemon is single
threaded, so the compromise for code clarity seems acceptable.