Add Associator class to update association records
When an association is modified in either Salesforce or the database, we
want to ensure that the same modification is applied in the other
system. Thus, when a Lookup is changed in Salesforce, we want to point
our database record to the synchronized record for that association, and
when a database record’s association is modified, we want to update the
Lookup in Salesforce.

To this end, we’ve added an Associator object, which handles the 
responsibility of identifying obsolete association links, and updating
those links in the appropriate systems.