Cache instances for query results during a run
We’ve introduced RunnerCache, an interstitial object which Runner 
consults to fetch its collections of Salesforce and database instances.
This abstracts out record fetching, and allows us to cache the instances
for the duration of a run, ensuring that we’re working with consistent
data.

While introducing this object, we’ve modified `#salesforce_instances` 
and `#database_instances` to return a generic Enumerator, rather than
forcing a specific block syntax on the classes which invoke these 
methods.