Redefine adapter layer for code clarity/accuracy
The current adapter implementation touches the attribute hashes a few
too many times. What we really want is a single layer which controls the
format of the information going into and out of the ActiveRecord object,
and a standard representation for the data elsewhere.

To that end, we’re standardizing on storing/passing information 
exclusively in a format understood by Salesforce, given that Salesforce
is the external system, and we have less/no control over it in code.

Redefining the format of our “normalized” Hash allows us to refactor our
`#convert` functionality down to a single method, and makes the code
much more understandable.