Restrict fields submitted through Salesforce API
At times it is beneficial to pull down fields from Salesforce which we
cannot push back up to Salesforce — read-only fields or formulas may
sometimes contain information which is valuable to synchronize into the
database.

In order to prevent conflicts in these situations, we can consult the
Salesforce API directly for information regarding whether or not a field
is read-only, and parse down our attributes accordingly.