Prevent unnecessary update calls to the API
If we _always_ submit an update request, we run into the issue of an
infinite update loop — the system modification timestamp for the record
in Salesforce gets bumped, and we pick it up again on the following run.

In order to get around this issue, we need to check the diff for any
changes, to determine whether or not an update is truly necessary.