Allow api_version to be specified in configuration
In order to support querying for recently-deleted records, we need to
bump the API version to 29.0. As an intermediate step (and to get our
VCR cassettes trivially compatible with the version change), we’re 
allowing users to specify an API version in their configuration, and
bumping the default up to our desired version.