Prevent deletion of records in “parallel” mappings
If, say, we have two mappings defined for the same combination of 
database and Salesforce record, and differentiate between them through
mutually exclusive `where` clauses, the current `Cleaner` behavior will
immediately delete our added records once we run it for the parallel
mapping (since the record is “invalid” for the other mapping).

The solution is to verify that a record is invalid for _all_ of the
mappings between any two record types before we proceed with our 
deletion.