Create top-level cache for new database records
It’s possible for a new database record under construction to have 
several inter-connected association records. In order to ensure that the
records we build maintain consistent references to the same in-memory
objects, we need to initialize our AssociationCache at the top level, 
and manually pass it in to each association’s build invocation.