Rescue from validation errors for specific records
When a single record fails to process within one of our task classes, we
should still continue to process the remaining records for that mapping.

Here, we add task-level exception logging, and rescue out of what are
generally harmless validation errors, to allow the task to continue its
processing for subsequent instances.