Prevent stale updates from being applied in sync
When changes are made to a record _while_ synchronization occurs, we
don’t want to synchronize the now-stale changes back in — this would
discard the more recent changes, introducing a race condition.

Instead, we can rely on the fact that the change in timestamp reflects
a change which will be picked up on the next synchronization tick. This
will _delay_ the record synchronization, but will ultimately result in
the most recent, most accurate data being propagated.