Expose problematic record IDs on sync exception
When a validation on either side of the integration fails, it’s 
extremely helpful to be able to locate the record which caused the sync
issue. To that end, we can expose the Salesforce ID (or database record
ID, in the event that an unsynchronized database record is causing the
problem) in our error message, to greatly assist in debugging efforts.