Add external update check to Salesforce records
In order to identify changes to Salesforce records made outside of the
Restforce::DB system, we need to pull down the ID of the user for which
the client is currently authorized, and compare it to the user who
last modified the record. This means pulling down the LastModifiedById,
which means rebuilding all of our cassettes, but sacrifices have to be
made sometimes.

Now that we can check for internal vs. external record updates, we can
wire that logic into our Synchronizer when determining whether or not
we should prevent the sync based on “stale” data.