Add `TimestampCache` to determine update accuracy
While we can tell which user _triggered_ the most recent changes to a
record in Salesforce, we can’t tell if any modifications to that record
were a result of a background Apex trigger or workflow (which apply any
changes as if they were the user whose actions initiated the callback).

In order to distinguish between updates made _by_ the worker and updates
made _in response to_ changes by the worker, we have to check the 
record’s update timestamp against the timestamp of the last known update
made by the system. `TimestampCache` serves as a mechanism to track the
values for this comparison.