Cache timestamps per-record type, not per-system
Since mappings can be set up one—to-many in terms of Salesforce models
mapped to database models, it’s dangerous to rely solely on the _system_
as an identifier in our timestamp cache. While Salesforce IDs are unique
across Salesforce record types, the same ID could correspond to multiple
database records, each of which should actually be treated as its own
Timestamp-able entity.