Defer worker setup until `#start` invocation
A lot of configuration was happening _within_ the initialization block
for Restforce::DB::Worker. Rather than have an object’s creation 
actively modify the global configuration for the library, we should 
defer all of our configuration until the worker is explicitly handed
control (via the `#start` method, which kicks off the polling loop).