Split out “safe” and “non-safe” reset methods
To make it easy for users of Restforce::DB to test its functionality
in a consistent way, we need to simplify the manner in which globally-
cached data and requests are reset.

The existing `Restforce::DB.reset` method has been repurposed for this
“safe” reset, while the more destructive reset functionality has been 
tucked into a new `Restforce::DB.reset!` method.