Prevent accidental duplication of unsynced records
Currently, when a record is created in Salesforce as part of the 
synchronization process, we don’t properly handle requests which time 
out or otherwise fail but were actually successfully submitted to 
Salesforce. This can result in the `#create!` call being repeated, and a
single database record becoming _two_ records in Salesforce. Obviously
not a great thing.

Salesforce supports an idempotent `upsert` action, however, which we can
use to create records without risk of accidentally duplicating them. The
caveat here is that it requires an external ID field to be defined for
the object in Salesforce.