Update restforce-db script with `meta` command
It’s kind of a pain for developers to have to manually add an external
Synchronization ID to their Salesforce models. To ease the difficulty
here, we’re exposing an additional command to automatically append the
field to a set of specified Salesforce models.

There are a lot of quirks to the implementation that I’ve attempted to
capture in both the script’s comments and its output.