Allow filtering of fields by purpose
In order to pull down the SynchronizationId for Salesforce records which
expose that field, we first need to be able to determine whether or not
the field is exposed at all for a given model.

Since `FieldProcessor` already fetches and caches various bits of 
information regarding field-level permissions, it makes sense to give it
the ability to tell us whether or not a field exists at all.