Pull down external IDs for Salesforce instances
In order to ensure that “upserted” Salesforce records are eventually 
attached to the database record that originally attempted to create
them, we need to pull down the `SynchronizationId__c` value with the 
rest of the fields for that instance.

Since Salesforce records aren’t required to have this field, we now
need to pass all proposed field lookups into our FieldProcessor to 
determine which are safe for consumption through the API.

This change means that Salesforce object descriptions will be requested 
and cached before ever querying for an object, which means regenerating 
all of our cassettes, unfortunately.