Introduce Attacher task to handle orphaned records
If an upsert request from the Initializer succeeds without properly 
linking the database record to the created Salesforce record (an
unfortunate situation which can happen any number of silly ways), we
need to identify the situation and properly link the Salesforce record
to the database record responsible for its creation.

The Attacher fills this role, identifying Salesforce records with 
lingering upsert IDs and locating their pair in the database.