Prevent inadvertent removal of modified records
There’s a subtle race condition in the Cleaner — in the event that a
record is updated between the query for “all” records and the query for
“valid” records, it will erroneously not be included in the list of 
valid IDs, despite still meeting the sync conditions.

In order to protect against this race condition, we can fetch the list 
of valid IDs _first_, ensuring that in the event of a race condition, we
err on the side of inclusivity.

We’ve also added some error logging around these situations, to call 
attention to those cases where a record _would_ have fallen prey to 
this dangerous race condition.