Prevent nil association lookups in Associator
In the event that a single lookup column is used for mutually-exclusive
`belongs_to` associations for a mapping, it’s dangerous to return a 
`nil` value to the `Associator`, as the _last_ association in this case
would get to call the shots, regardless of which association is 
populated.

Instead, we should simply ignore missing associations. De-assigning
records is a much less frequent interaction than re-assigning them, so
we should optimize for that case.