Handle multiple associated target mappings
In the case where a model has multiple (mutually exclusive) mappings
defined, it’s possible that a HasOne or HasMany association needs to
traverse the inverse associations on more than one of those mappings to
identify the correct associated record.

To that end, we cannot assume that the _first_ mapping is the correct 
one, and must instead iterate through all mappings which expose a valid
inverse association until we either find one which gives us a result, or
run out of mappings to investigate.