Enable Faraday::Request::Retry middleware
Running into an occasional request timeout is problematic, as it 
generally means the changed record will _not_ go through the proper
synchronization process. In order to cut down on the impact of errant 
latency on our sync, we can enable Faraday’s built-in `Request::Retry`
middleware to ensure that idempotent requests are re-attempted in the
event of a timeout.