Add restforce:populate rake task
As a project continues to grow, it may often be necessary to add new
fields to one or more existing mappings. To ease the tedium of 
constantly resetting the database or manually populating information,
we can introduce a rake task to perform a one-time initial seed of data
for individual fields.