Double-check Cleaner’s proposed invalid record IDs
The working theory is that in the event of one of Salesforce’s servers 
having a slow clock, it’s possible for a bulk update to send a record 
“back in time”, causing it to show up in the results for a query run 
against an earlier SystemModStamp. In this event, the record could show
up in the query for “all” records, but not in the query for “valid” 
records, and thus the Cleaner would attempt to cull that record.

To work around this potential edge case, we can essentially check the
Cleaner’s work, asserting that each proposed record has indeed been
legitimately invalidated before deleting them from the database.

The overhead we’re introducing here should be tolerable — the frequency
with which records are invalidated will generally be low, and the side-
effects of inadvertently deleting a record are generally steep enough to
justify the additional caution.