Refactor Worker to run sync within subprocesses
In an effort to make the long-lived sync process more tenable, we need
to ensure that Ruby’s tendency to consume large amounts of memory (then 
never free it back up) is mitigated.

To this end, we can ensure that memory temporarily consumed during the 
sync is freed back up by performing all of our sync tasks within a 
subprocess. When the process exits: free garbage collection.

In order to get this working, we also needed to invoke the appropriate 
before/after hooks for file reopening, which, as a side effect, made it
trivial to implement the USR1 signal for log rotation.

TODO: The timestamp and field info caches are going to be a bit busted
after this first pass. I’ll be cleaning those up in separate commits.