Handle Restforce::AuthenticationError bubbling
There’s an underlying “bug” in the Restforce library, which results in
an error when the response from Faraday doesn’t have a body for some
reason. It’s still unclear what the conditions leading to this edge case
are, but it seems tied to a number of other thread-safety issues, and
we should generally just take it as our cue to error out of the current 
fork and try everything again.