Handle mutually-exclusive mappings on forced sync
In the event that a model has two mutually-exclusive mappings to 
Salesforce, each record will always correspond to only one of them —
this results in an error when we attempt to work with the Salesforce
instance for the exclusive mapping.

The solution is just to skip these — we can generally assume that a 
record with a Salesforce ID corresponds to _something_ in Salesforce.