Reduce logging level to :warn for auth errors
The "Request was closed prematurely" exception is triggered fairly
frequently during synchronization worker loops, and doesn’t seem to
represent anything actionable. Since we handle it at the sync level by
retrying the entire processing loop, it should be sufficient to log the
exception at the :warn level and proceed as usual.